requirements.txt
fastapi==0.127.1
uvicorn==0.40.0
pydantic>=2.7.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
websockets==12.0
